#0 | Removing dp core dependency in Usercacheupdaterv2 flink job
