+++ jobs-core/src/main/scala/org/sunbird/job/cache/DataCache.scala
@@ -182,6 +182,22 @@ class DataCache(val config: BaseJobConfig, val redisConnect: RedisConnect, val d
     this.redisConnection.del(key)
   }
 
+  def hdel(key: String, fieldSeq: Seq[String]): Unit = {
+    this.redisConnection.hdel(key, fieldSeq: _*)
+  }
+
+  def hdelWithRetry(key: String, fieldSeq: Seq[String]): Unit = {
+    try {
+      hdel(key, fieldSeq)
+    } catch {
+      case ex: JedisException =>
+        logger.error("Exception when deleting fields in hash", ex)
+        this.redisConnection.close()
+        this.redisConnection = redisConnect.getConnection(dbIndex)
+        hdel(key, fieldSeq)
+    }
+  }
+
 }
 
 // $COVERAGE-ON$
