+++ user-org-jobs/user-cache-updater-2.0/src/main/scala/org/sunbird/job/usercache/task/UserCacheUpdaterConfigV2.scala
@@ -1,12 +1,12 @@
-package org.sunbird.dp.usercache.task
+package org.sunbird.job.usercache.task
 
 import com.typesafe.config.Config
 import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.api.java.typeutils.TypeExtractor
-import org.sunbird.dp.core.job.BaseJobConfig
-import org.sunbird.dp.usercache.domain.Event
-import scala.collection.JavaConverters._
+import org.sunbird.job.usercache.domain.Event
+import org.sunbird.job.BaseJobConfig
 
+import scala.collection.JavaConverters._
 import java.util.{List => JList}
 
 class UserCacheUpdaterConfigV2(override val config: Config) extends BaseJobConfig(config, "UserCacheUpdaterJobV2") {
