+++ user-org-jobs/user-cache-updater-2.0/src/test/scala/org/sunbird/job/fixture/EventFixture.scala
@@ -1,4 +1,4 @@
-package org.sunbird.dp.fixture
+package org.sunbird.job.fixture
 
 object EventFixture {
